Improve reverse proxy "create" usability
In the "create" scenario, some form fields are shown/hidden depending on
the first character typed in the Name field. If fields are initially
hidden, it is not clear what to do with Advanced settings.

As the "/" leading character to build a path rule is less common
(type=ProxyPass) and requires fewer fields than a host name matching
rule (type=VhostReverse) the PR initially displays the
form state for VhostReverse, with all available fields displayed.